part2

+++ src/main/java/facades/PersonFacade.java
@@ -8,6 +8,8 @@ package facades;
 import dtos.PersonDTO;
 import dtos.PersonsDTO;
 import entities.Person;
+import exceptions.MissingInputException;
+import exceptions.PersonNotFoundException;
 import java.util.ArrayList;
 import java.util.List;
 import javax.persistence.EntityManager;
@@ -34,8 +36,11 @@ public class PersonFacade implements IPersonFacade {
     
 
     @Override
-    public PersonDTO addPerson(String fName, String lName, String phone) {
+    public PersonDTO addPerson(String fName, String lName, String phone) throws MissingInputException {
         EntityManager em = emf.createEntityManager();
+        if(fName.length() == 0 || lName.length() == 0){
+            throw new MissingInputException("First Name and/or Last Name is missing");
+        }
         Person person = new Person(fName, lName, phone);
         try {
             em.getTransaction().begin();
@@ -50,32 +55,40 @@ public class PersonFacade implements IPersonFacade {
     }
 
     @Override
-    public PersonDTO deletePerson(int id) {
+    public PersonDTO deletePerson(int id) throws PersonNotFoundException {
         EntityManager em = emf.createEntityManager();
-        Person person = em.find(Person.class, id);
         try{
+            Person person = em.find(Person.class, id);
+            if(person == null){
+                throw new PersonNotFoundException("Could not delete, provided id does not exist");
+            }
             em.getTransaction().begin();
             em.remove(person);
             em.getTransaction().commit();
             return new PersonDTO(person);
+            
         }finally{
             em.close();
         }
     }
 
     @Override
-    public PersonDTO getPerson(int id) {
+    public PersonDTO getPerson(int id) throws PersonNotFoundException{
         EntityManager em = emf.createEntityManager();
         try {
             Person person = em.find(Person.class, id);
+            if (person == null){
+                throw new PersonNotFoundException("No person with provided id found");
+            }
             return new PersonDTO(person);
+            
         } finally {
             em.close();
         }
     }
 
     @Override
-    public PersonsDTO getAllPersons() {
+    public PersonsDTO getAllPersons(){
         EntityManager em = emf.createEntityManager();
         TypedQuery<Person> query =  em.createQuery("SELECT p FROM Person p",Person.class);
         List<Person> personList = query.getResultList();
@@ -84,11 +97,17 @@ public class PersonFacade implements IPersonFacade {
     }
 
     @Override
-    public PersonDTO editPerson(PersonDTO p) {
+    public PersonDTO editPerson(PersonDTO p) throws PersonNotFoundException, MissingInputException{
         EntityManager em = emf.createEntityManager();
+        if(p.getfName().length() == 0 || p.getlName().length() == 0){
+            throw new MissingInputException("First Name and/or Last Name is missing");
+        }        
+        Person person = em.find(Person.class, p.getId());
+        if(person == null){
+            throw new PersonNotFoundException("No person with provided id found");
+        }
         try{
             em.getTransaction().begin();
-            Person person = em.find(Person.class, p.getId());
             person.setFirstName(p.getfName());
             person.setLastName(p.getlName());
             person.setPhone(p.getPhone());
